tests/wpt/web-platform-tests/media-feeds/link-rel-media-feed.html

<!DOCTYPE html>
<title>Test that "media-feed" is a supported value for the `rel` of a `link`</title>
<link rel="help" href="https://wicg.github.io/media-feeds/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    const result = document.createElement("link").relList.supports("media-feed")
    assert_true(result, "Expected true if media-feed is supported as a link relationship");
}, 'link element supports a rel value of "media-feed".');
</script>
